Further optimize equipment screen
Updating then rebuilding for each change is redundant.

Calling Rebuild for OnEquipmentChanged was not necessary.
If the slot id is the same then updating and calling rebuild was not necessary.
If both slots are weapons, calling rebuild was not necessary.

Move weaponslot list to a static variable to not be recreated every build.