Allow players to change the Race of Troops
Puts a button next to Culture that changes the race/monster/species of the troops if they have mods that add humanoid monsters.

Currently the button and "Race" header show regardless of if they have available alternate races, but I left this so it shows that the feature is available rather than it being a hidden thing.

I couldn't figure out how to get the panel to be perfectly centered, if you want to get it more centered then please :). When switching to a race with a long name, itll push the Culture side out of the way rather than moving over to be perfectly centered.

I also left the translations null since I do not know what translator you use.

Races I tested with are from The Old Realms. Not exactly on The Old Realms, but just the races.