Fix & Update Volunteer Swapping Behavior
An issue was caused by caching troop matches. When the player visited a settlement and swapped troops, there was no actual logic to swap them back, since it relied on backward matching. I had only set up the matching cache for the player rather than for the AI. This led to AI Lords being able to recruit Player Troops regardless of player config.

Instead of extending the cache to the AI, I realized that doing so would cause improper matches, because mods with multiple troop trees or other variations would not accurately match what they had before. So, I decided to save what the volunteers were before the swap, then swap them back when the player leaves the settlement, begins waiting in it, or saves the game (to ensure the volunteers swap back before leaving game or crashing). This ensures that the original troops do not get swapped improperly. Please check if my logic here is correct, as I am not completely sure how recruiting and your system work.

I removed the VolunteerSwap file as it is no longer needed and interferes with this method. That also reduces the amount of background calculations.

I also updated Settlement to use the cached Player.Clan or Player.Kingdom, since without it, the new swapping method did not function correctly.