Fixes issue with Elite Retinue in Warlord's Battlefield
Not sure what the actual cause/reason is but in Warlord's Battlefield, the elite retinue couldn't be ranked up and gave the popup saying that House Champion cannot outrank House Champion. This changes the check to only consider the basic retinue, as the elite retinue doesn't need to be checked regardless.